Fix compile error on some platforms

+++ src/tracing/json/json_tracing.cpp
@@ -348,7 +348,7 @@ void JsonBackend::LogMessageSend(MessageSendInfo & info)
     DecodePayloadHeader(value["payloadHeader"], info.payloadHeader);
     DecodePacketHeader(value["packetHeader"], info.packetHeader);
     DecodePayloadData(value["payload"], info.payload, info.payloadHeader->GetProtocolID(), info.payloadHeader->GetMessageType());
-    value["messageTotalSize"] = info.messageTotalSize;
+    value["messageTotalSize"] = static_cast<int>(info.messageTotalSize);
 
     OutputValue(value);
 }
@@ -375,7 +375,7 @@ void JsonBackend::LogMessageReceived(MessageReceivedInfo & info)
     DecodePayloadHeader(value["payloadHeader"], info.payloadHeader);
     DecodePacketHeader(value["packetHeader"], info.packetHeader);
     DecodePayloadData(value["payload"], info.payload, info.payloadHeader->GetProtocolID(), info.payloadHeader->GetMessageType());
-    value["messageTotalSize"] = info.messageTotalSize;
+    value["messageTotalSize"] = static_cast<int>(info.messageTotalSize);
 
     OutputValue(value);
 }
